src/requirements.txt
pycti==4.5.4
PyYAML==5.4.1
pysftp>=0.2.9